pyproject.toml
[project]
name = "vehiclemodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "datetime>=6.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pint>=0.25.2",
    "pyright>=1.1.407",
    "pyyaml>=6.0.3",
    "ruff>=0.14.8",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
pythonPlatform = "Darwin"
typeCheckingMode = "basic"
